Implemented comment's feature.

+++ test/integration/posts/show_post_test.rb
@@ -10,7 +10,7 @@ class ShowPostTest < PostsTestBase
     assert page.find('h3', text: /^#{@post.title}$/)
     assert page.find('p', text: /^#{@post.content}$/)
     assert page.find('#post-author', text: /#{@post.user.name}/)
-    assert page.find('#post-calendar', text: /#{I18n.l(@post.created_at, format: :short)}/)
+    assert page.find('#post-calendar', text: /#{I18n.l(@post.created_at, format: :long)}/)
   end
 
   test "go back from show post" do
